make code of homeDir, configFile & configDir not os specific anymore
