+++ packer/config_file.go
@@ -1,7 +1,9 @@
 package packer
 
 import (
+	"log"
 	"os"
+	"os/user"
 	"path/filepath"
 )
 
@@ -37,4 +39,48 @@ func ConfigTmpDir() (string, error) {
 		return "", err
 	}
 	return td, nil
+}
+
+func homeDir() (string, error) {
+
+	// First prefer the HOME environmental variable
+	if home := os.Getenv("HOME"); home != "" {
+		log.Printf("Detected home directory from env var: %s", home)
+		return home, nil
+	}
+
+	// Fall back to the passwd database if not found which follows
+	// the same semantics as bourne shell
+	u, err := user.Current()
+
+	// Get homedir from specified username
+	// if it is set and different than what we have
+	if username := os.Getenv("USER"); username != "" && err == nil && u.Username != username {
+		u, err = user.Lookup(username)
+	}
+
+	// Fail if we were unable to read the record
+	if err != nil {
+		return "", err
+	}
+
+	return u.HomeDir, nil
+}
+
+func configFile() (string, error) {
+	dir, err := homeDir()
+	if err != nil {
+		return "", err
+	}
+
+	return filepath.Join(dir, defaultConfigFile), nil
+}
+
+func configDir() (string, error) {
+	dir, err := homeDir()
+	if err != nil {
+		return "", err
+	}
+
+	return filepath.Join(dir, defaultConfigDir), nil
 }
